+++ integration_test/button_test.dart
@@ -1,8 +1,8 @@
 
 import 'package:flutter/material.dart';
+import 'package:flutter_recipe_app/presentation/component/button/big_button.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:integration_test/integration_test.dart';
-import 'package:flutter_recipe_app/component/button/big_button.dart';
 
 void main() {
