+++ lib/presentation/component/button/medium_button.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_recipe_app/ui/app_color.dart';
+import 'package:flutter_recipe_app/presentation/ui/app_color.dart';
 
 class MediumButton extends StatelessWidget {
   final String buttonText;
