+++ lib/presentation/component/dialog/rating_dialog.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_recipe_app/component/button/small_button.dart';
-import 'package:flutter_recipe_app/ui/app_color.dart';
+import 'package:flutter_recipe_app/presentation/component/button/small_button.dart';
+import 'package:flutter_recipe_app/presentation/ui/app_color.dart';
 
 class RatingDialog extends StatefulWidget {
   final String title;
