+++ lib/presentation/screen/home/splash_screen.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_recipe_app/component/button/medium_button.dart';
-import 'package:flutter_recipe_app/ui/app_color.dart';
+import 'package:flutter_recipe_app/presentation/component/button/medium_button.dart';
+import 'package:flutter_recipe_app/presentation/ui/app_color.dart';
 
 class SplashScreen extends StatelessWidget {
   const SplashScreen({super.key});
